Supress numpy VisibleDeprecationWarning

+++ digits/utils/image.py
@@ -105,7 +105,7 @@ def upscale(image, ratio):
     channels = image.shape[2]
     out = np.ndarray((height, width, channels),dtype=np.uint8)
     for x, y in np.ndindex((width,height)):
-        out[y,x] = image[math.floor(y/ratio), math.floor(x/ratio)]
+        out[y,x] = image[int(math.floor(y/ratio)), int(math.floor(x/ratio))]
     return out
